+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -22,3 +22,4 @@
 * Provides LifecycleOwner implementation for Activities that use the plugin that do not implement it themselves.
 * Implements retrieval of camera information.
 * Updates README.md with plugin overview and adds contribution guide to CONTRIBUTING.md.
+* Implements video capture.